General update of the API for signup, login; created new folder utils and moved functions for bcrypt there.
In login I deleted the check of JWT_SECRET, all of this kind of errors will be catched by the last error message, becuase user doesn't need to know what is the problem with the server. Additionally, I fixed the issue with the password names (made them according to the schema.prisma. One more update is a new API point -  logout. Although we are using tokens, this message can be displayed when user logs out.