Added comments, blogs, voting system
Updated prisma schema with the Vote model. Also added upvotes and downvotes for BlogPost and Comment models. If the person voted for the comment already, subsequent request to vote for it will result in removing of the vote from the comment/blogpost. Thus, every user can vote for the comment/blogpost only one time. If they removed the vote, they can vote again for the same comment. blogs and comments API endpoints are implemented. Have yet to  implement the BlogPost-Comment sorting.